requirements.txt
scrapy==2.11.2
pyspellchecker==0.8.1
jinja2==3.1.4